The JTable in the Dashboard is now interactable
The JTable in the Dashboard has a mouse listener that basically
makes the rows act like buttons, if a certain row is clicked, it will
ask the user about the quantity of the certain transaction item, and
then puts it in the itemsArea JTextArea. The JTextArea shows the buy and
sell subtotals for each transactionItem and the buy and sell for the
entire transaction.

Some other stuff were fixed like in the MaterialDAO, I forgot to update
the SQL query itself which causes SQL syntax errors