Record a buy and sell buttons now have functionality!
The Controller methods of the buy and sell buttons basically just
fills in the missing fields of the transaction and transactionItems when
they were initially constructed when choosing from the materialsTable,
then some other validations like checking if one of the transaction
items quantity is greater than the stock when doing the record sell
button are done, THEN the Transaction and TransactionItems are finalized
and the other GUI components in the dashboard gets changed like the
stock quantities of the chosen transactionItems.

I also changed the DAO methods by enclosing the ResultSet objects inside
a try-with block to avoid memory leaks and stuff.