I added an append material button in the admin
I also polished the system a bit, like when a material is chosen from
the Dashboard, and there is an invalid input, it doesn't completely
reset the prompt, but it repeats where the user did a mistake. I also
fixed a bug where the totalAmount wont reset everytime a user finalizes
a transaction, I fixed it by setting totalAmount to 0.0 after a buy or
sell is recorded, I also made it so that when a user does some queries
like the getAllTransactions, if no row is returned, outputArea will say
that no rows were returned etc2.